hwopus: Document second member of HwopusHeader
This second member indicates the Opus encoder's entropy coder's final
range. While nothing in the service actually does anything with the
final range as far as I can tell, this is specified within RFC 6716
(https://tools.ietf.org/html/rfc6716, see section "6. Conformance") as a
means of ensuring decoder conformance. States as follows:

"In addition, a compliant decoder implementation MUST have
 the same final range decoder state as that of the reference decoder."

So what is likely done when performing compliance testing is the data is
encoded, and then the final range of the encoder is sent via this
header, then during decoding, the final range would be checked against
to ensure that it's valid.